+++ src/wp-admin/includes/class-rollback-auto-update.php
@@ -202,8 +202,8 @@ class WP_Rollback_Auto_Update {
 		sleep( 2 );
 
 		$this->restart_updates();
-		$this->restart_core_updates();
-		$this->send_update_result_email();
+		//$this->restart_core_updates();
+		//$this->send_update_result_email();
 	}
 
 	/**
@@ -289,6 +289,8 @@ class WP_Rollback_Auto_Update {
 		$skin     = new \Automatic_Upgrader_Skin();
 		$upgrader = new \Plugin_Upgrader( $skin );
 		$upgrader->bulk_upgrade( $remaining_auto_updates );
+		$this->restart_core_updates();
+		$this->send_update_result_email();
 	}
 
 	/**