+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 [unreleased]
 * update to remove static init hook and load `class-rollback-auto-update.php` in `class-wp-upgrader.php`
+* try to avoid sending update email twice
 
 #### 4.1.2 / 2023-01-25
 * update `move_dir()` for new parameter