+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 [unreleased]
 * during `WP_Rollback_Auto_Update::restart_updates` remove shutdown hook for `WP_Upgrader::delete_temp_backup`
+* skip second sequential call to `create_backup`
 
 #### 4.1.2 / 2023-01-25
 * update `move_dir()` for new parameter